pyproject.toml
[tool.poetry]
name = "git-id"
version = "0.1.0"
description = "An ID manager for git"
authors = ["Yuichiro Smith <contact@yu-smith.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
click = "^7.1.2"
pyyaml = "^5.3.1"
gitpython = "^3.1.9"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
